feat: Continuous reading of values
move encoding stuff to own class

Issue #6
